+++ changelog.md
@@ -12,6 +12,7 @@
         }   
       }
     ```
+* Potions duration is now a range Min Max value instead of a fixed value
 
 ## Alpha 3.4.1
 * Added TiCon Attack config option. When enabled, mobs will be able to use some of the Tinker tools modifiers when attacking (e.g. Sharpness, Fiery, Conducting when on fire, etc.)